fix: rs-sdk with no features does not build
