+++ packages/rs-sdk/src/mock/noop.rs
@@ -1,0 +1,5 @@
+pub trait MockRequest {}
+impl<T> MockRequest for T {}
+
+pub trait MockResponse {}
+impl<T> MockResponse for T {}
